ADDED:  AddContactMech component (resuable, in principle, still needs some work) ADDED:  CreateContact page added and tested.
Notes:  CreateContact adds a new Contact’s personal info and contact mechanisms to the db.  Since I focused on getting the core operation into place (and wanting to make some form of AddContactMech available ASAP to the team), I have not yet addressed a number of issues that CreateContact and AddContactMech still have:

(1) Margins within the CreateContact form, the submit button placement, other appearance stuff need tweaking/harmonizing with other pages.

(2) Pull-down menus have not yet been added to AddContactMech (or AddPersonalInfo) child pages. For now all fields are input type=“text” only, which I did just so I could reach the point of having something working all the way through to the db.  Now that I  have confirmed all the contact mechanisms reach the DB, I will try to amend AddContactMech ASAP to have drop-down menus.

(3) No error handling or validation of data is included yet—- right now the form will let anything be submitted, so I have only submitted values I know will all validate. Since validation is done by the back-end controller layer, I am not sure what/how much validation I need to be doing on the CreateContact page itself.

(4) After successful submission of the form, the re-direct to MyContactsPage is not working.  That is, upon clicking Create Contact to submit the form, it redirects to the home page instead of to the MyContacts page as planned.  The problem looks fixable, some detail with React Router I will try to figure out soon…

(5) CreateContact is not yet using any of the reusable components others on the team have made, e.g., Submit Button.  Now that the basic functionality is in place, I hope to explore and use reusable components.

(6) AddContactMech and AddPersonalInfo still have a ton of repetition of HTML in them and could benefit from themselves pulling in a reusable child component for text fields, as done in the Pluralsight video and as others have probably already explored.

(7) After successful posting to the db, the CreateContact page should be receiving the newly generated partyId.  It WAS working, but I must have changed something small when cleaning up, because now the state of addedContactPartyId is not updated and console gives this error:  “Warning: setState(...): Can only update a mounted or mounting component. This usually means you called setState() on an unmounted component. This is a no-op. Please check the code for the CreateContactPage component.”  Note:  getting the new partyId back to the CreateContact page is not vital to the page’s function, which is done once validatable data is sent out.  Still, I would like to close the loop by getting the new partyId back to the page from the Store after a successful ajax call.  Even though the new partyId won’t be rendered on the CreateContact page (which the user is redirected away from, if succesful), maybe it matters for error checking or possibly adding a pop-up message to report the number to the user.