UPDATED:  AddContactMech page now has working pull-downs for State/Province and Country.  AddContactMech and its child pages StateProvinceOption and CountryOption have been moved into views/components/common/
Note to Lucas:  I attempted to incorporate the common SubmitButton on the AddContactForm page but could not figure out how to pass down the onClick property I needed to get it to work.  I put a comment at the bottom of SubmitButton.js and returned the <button> tag to your original code.  Maybe we can discuss soon and I can try again?