UPDATED:  FindContacts page now uses jQuery-ui tabs instead of accordion.
Note to Lucas (in particular):  I switched from accordion to tabs; as you pointed out, it makes the two kinds of search types more obvious to the user, and so I think tabs are a better choice than accordion.  (I commented out the accordion-related HTML and the jQuery command for the accordion but did not completely delete it for now.  If the consensus continues to favor tabs over accordion, eventually I'll delete the commented out accordion-related code.)